Fix scriptlet error when built without firewalld zone support
In 5cbc0451, Daniel added a use of the %firewalld_reload macro
that's conditional on `with_firewalld`, but made the BuildRequire
on firewalld-filesystem (which provides the macro) conditional
on `with_firewalld_zone`. So if you build without firewalld_zone
but with firewalld, the macro is not expanded and the built
package has a broken %post script. This fixes it by making the
BuildRequire also conditional on `with_firewalld`.